refactor: Judge의 역할을  Dealer로 옮김
